[Jobs,Workflows] Duplicated requests
